temporary fix replication restart when remote is down
quick hack to fix the case when remote is down for a while but we still
need to keep it alive.

This is done by increasing the restart intensity before the supervisor
give up (do it the right way this time). Also fix a bad pattern
matching.

This fix will be superceded by laer monitorning the node and pause the
replication instead of crashing it like we do now.