new barrel_rpc library
this  is tailored rpc protocol for barrel. This is dedicated to the replication as for now it's not expected to use the protocol for direct usage of the remote. All queries should happen on the local node or upper node of the chain that replicated the data.

This module is simpler than previous one as it's only talking via the Erlang distribution for now. Support for an extended distribution will happen.

Supported functions:

* update_docs: stream docs to the remote (we send / doc)
* fetch_docs: stream docs from remote
* subscribe/unsubscribe remote changes
* revsdiff
* system docs operations