add disk and memory monitoring
raises an alarm when we go over a memory or disk limit. Instead of using the os_mon application we are using modules imported from rabbitmq for that so we will be able to distinct the folders we really want to monitor.

I will be happy to revisit that later to see if simply using os_mon is enough.