add new pull api
the pull api allows you to retrieve all Ids and their history based
on a path. It also allows you to continue from the last bound.
the format is similar to a changes feed so you can use it to replicate
the docs locally based on a path.

also while i'm here remove the current limit of the fold function.