simplify the node checker
simpler code so the module isn't linked to the replication job. the difference with `erlang:monitor_node/2` is that the server will keep monitor the remote node instead of doing it once for the process.

Also when the remote node comes down, it start to ping it at interval until it come online. This allows the process to wait for this node until it comes offline.the