try to restart the replication if applying a change fail
workaround to not crash if something happen while storing the change
on the replication target. Instead we log the error, stop the changes
feed, sleep 1 second and wait for the source.

This is a workaround, proper fix would be
breaking the circuit . This will come in another commit, but for now
unblock the situation.

fix KOB-60331 and KOB-60332