improve job handling
this is the first pass in the way we are handling jobs on the vm.
Instead of starting them through the db process, we directly try to
fetch an available job from the pool.

If no jobs are available we return `{error, dropped}` . If not we
execute the job in the pool.

TODO:

    - retrieve a context to pass to the storage when executing the job
in the. It will allows us to pass a Connection to a remote database or
snapshot of the local database for consistent reads.
    - improve job pool to not recycle jobs, so we can have a pristine
process
    - maybe add the write to the jobs.