fix stream agent monitoring
make sure to demonitor a stream agent when to the manager: we were
checking against the PID and not {Stream, Stream} ref witch is the
monitor value.

ALso handle the case when a stream agent comes down and the stream
was unsubscribed. In such case the stream informatioin has been removed
but we kept a monitor around due to the agent running.