use Serializable Snapshot Isolation to store the doc
until now we were mutating documents in a single process. This change mutate the document directly in clients process (without message passing) and only pas the final batch to the sequencerr process to append them in order in the database.

This will bring many advantage like allowing us to improve the indexation speed and conflicts return.