remove ARS related data when indexing
main goal of the replication is to accept docs and merge them
consistently and let the user or the current storage create its own
views as its on pace over the changes feed.

With this change we are removing doc tagging on merge which was slowing
the writes too much for our main goal.