refactor: simplify data storage
with this changes, merging the revision tree and appending a sequence to
documents happen inside one gen_server. Documents are sent by batch to
the `barrel_server` process and then commited to the doc. this makes the
storage simpler to reason. caveat: fetching the metadata and merge for
all docs are serialized inside one gen_server. On the other hands
fetching the metadata should be fast. Reads and queries are done in //.