change sequence to a tuple {Epoch, Seq}
this change prepare the future distribution of barrel by replacing the
simple monotonic counter by a pair of {Epoch, Seq} whhere the epochis
monotically increasing on start and the sequenceis a counter
increemented on each changes.

This allows us to make sure the sequence of changes will constantely
increasing event in the presence of partial data on restart. It also
make the sequnce independant of the sotorage.

The epoch store is a configurable module set at startup. It is
repsonsible of incrementing and giving an epoch to each barrelsls. By
default the epoch stores is local and the eopch is a timestamp
corresponding to the start time. Dialoging with the epoch store is done
via message passing.