return faster when waiting a view
this changes allows the user to ask for a view refresh from the last
time it got an update from it.

expected result: allows the user to reuse a session and not wait for
uselesss updates. (improve the read your own writes case).

new behaviour:

idle state:

* if last updated sequence =< last sequence indexed it returns immediately,
else it pospone the events
* hen a new db update event is received it process the changes. If any new
changes it goes to next state indexing. Else it keeps the current state

indexing state

* on refresh request, if last sequence given < minimum sequence currently
processed it returns immediately
* on refresh requests if sequence give is in th emap, just add the
requester to the list of waiters for this sequence. Else wait for the
last received
* in other case wait for the last receive