replace epoch by timestamp
This change allows the user to use a custom timestamp when storing the
data. This allows the application to use its own way to pass a timestamp
whn storing a change.

When a timestamp is given the sequence returned in the change document
is formatted as "<timestamp>-<sequence>".

A method allowing the user to retrieve the changes based on the
timestamp will be added.